# salesman-ecommerce/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/kpi/list": {
            "get": {
                "tags": [
                    "KPI"
                ],
                "summary": "Get list all kpi",
                "operationId": "b3bc2e3e5f248728d03d99b134874e8f",
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/user/kpi": {
            "get": {
                "tags": [
                    "KPI"
                ],
                "summary": "Show kpi of salesman",
                "operationId": "98de794aad73ed31cd914ec678454b27",
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/admin/kpi/save": {
            "post": {
                "tags": [
                    "KPI"
                ],
                "summary": "Add kpi",
                "operationId": "d018f806b44855a6260f39665e90501d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "salesman_id": {
                                        "type": "string",
                                        "example": "USER00000003"
                                    },
                                    "order_amount": {
                                        "type": "integer",
                                        "example": "10"
                                    },
                                    "checkin": {
                                        "type": "integer",
                                        "example": "10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                }
            }
        },
        "/api/admin/kpi/delete/{kpi_id}": {
            "post": {
                "tags": [
                    "KPI"
                ],
                "summary": "Delete KPI",
                "operationId": "00cf32e21cba8c41f41a27aff00badf8",
                "parameters": [
                    {
                        "name": "kpi_id",
                        "in": "path",
                        "description": "kpi_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}